[RTTI] Handle cxxabi vtables being referenced via RELOC_COPY
32-bit ELF binaries that are dynamically linked to the C++ runtime may
use a copy relocation for the vtable. The vtable itself will be defined
in the `.bss` section, and the copy relocation will cause the dynamic
linker to populate it at load time from the C++ runtime library.

Detect this by looking for a symbol pointing to the start of the vtable data,
two pointers before the vtable address.