[RTTI] Handle type names emitted by GCC with a leading `*`
GCC emits a leading `*` to indicate that the type info is internal and
its name can be compared via pointer equality. It is not part of the
type name. This was only being handled when followed with `N`, but it
can apply to any mangled name.

Additionally, this updates some `std::string::find(...) == 0` calls in
the adjacent code to use `std::string::rfind(..., 0) == 0` as that bails
out of the string comparison as soon as the prefix does not match,
rather than continuing to search the entire string.