[MachO] Don't add symbols when processing bind fixups
The target of the bind opcode has its symbol applied when the dynamic
symbol table is parsed.

This was previously adding a symbol at the fixed-up address with the
name of the symbol being bound. This was resulting in dozens of
`_OBJC_CLASS_$_NSObject`, `_OBJC_METACLASS_$_NSObject`, and
`__objc_empty_cache` symbols being created: one for each `objc_class`
that referenced those symbols.