Fix clean builds when CMake is using Unix Makefiles
The `$<TARGET_FILE:binaryninjaapi>` syntax does not do what is intended
for Unix Makefiles.

In practice there's no reason for this custom command to depend on the
`binaryninjaapi` target. The Rust API uses core's C APIs directly rather
than via the C++ API, and it already depends on binaryninjacore.h via
`RUST_API_SOURCES`. Since it's running `cargo check` and not doing a
full compilation this header-only dependency is sufficient.