[WARP] Fix very large binaries from being prevented from creating WARP files of all the functions
This was caused by the internal flatbuffer verifier reaching the max number of tables, this is fine and the solution is to split the large chunk into smaller ones, effectively creating more flatbuffer "views" of tables.

This is fine from a performance perspective because we already have optimized for many small chunks (e.g. we have a lookup table stored next to each chunk).

I left some comments for future improvements but this should be good now.

Also alongside this change we improved generating performance by 2x on some pathological binaries.